Note
I think the problem is that I am changing the estimation problem when I am changing the number of samples used within the AICc-formula as the distances etc are computed to more training points. Still don't understand why the sd increases in this case though. And why the right amount of data should give such a small sigma. Is that smart here? Check that out. Note also that optimization over several test samples is not doable by now as test samples need to be handled separetly (currently not implemented and would be hard computationally).