Adds script comparing to Lundberg (#88)
Adds script under inst comparing our method to the Python shap package of Lundberg. Output from the script will eventually be put into the vignette, but would be good to have it as script here as well. The script includes timing of the explanation relevant code. We are about 10% faster.

Also added the xgboost binary file loaded by the Python script.

Fixes #59.