Fixed visual bug in `plot.shapr()` which gave extra whitespace in the strings "feature = value".
My proposed solution is to add trim whitespace as we previously got extra white space for those features that has different number of digits. E.g., if a feature is "1000" for one observation and "5" for another then we would previously get the strings "1000" and "   5", i.e., the latter/smaller one got three extra white spaces.