Fix linting and stylr (#364)
* Fixed visual bug in `plot.shapr()` which gave extra whitespace in the strings "feature = value".

My proposed solution is to add trim whitespace as we previously got extra white space for those features that has different number of digits. E.g., if a feature is "1000" for one observation and "5" for another then we would previously get the strings "1000" and "   5", i.e., the latter/smaller one got three extra white spaces.

* Added script which describes and illustrates the visual bug in `shapr.plot()`. The visual bug is not present after applying the bugfix.

* update vdiffr test files

* remove "base::" for trimws()

* package man

* stylr

* disable cyclocomp and indentation linters

as indentation don't agree with rstudio shortcuts for data.table and other edge cases. May enable it in the future. Also, too cumbersome to avoid the cyclocomp linting complaints so may as well ignore that too.

---------

Co-authored-by: LHBO <lholsen@math.uio.no>