Not 100% sure this actually works as it should
Nope, the bootstrapping uses features and n_features, so not using the groups. Can I make it generally applicable? with coal+ n_per_coal or so everywhere?  Discuss this with Lars.
Also add a test where the groups corresponds to one feature each as a sanity check of wethter i can identical results.